Implement running of workflows in Automate API
Extended the workflow service to include Automate API calls to initiate workflows, as well as incorporated exception handling for these operations. Also, added a new endpoint in the Automate API router to respond to these calls and updated some dependencies in the same file. Environment variables are now used in Workflow Service for better manageability.